Move jwt stuff to utils
